+++ src/main/java/com/crs/main/model/Booking.java
@@ -1,8 +1,6 @@
 package com.crs.main.model;
 
 import com.crs.main.util.enums.BookingUtils;
-import com.crs.main.util.enums.PaymentUtils;
-import com.crs.main.util.enums.RentUtils;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -29,8 +27,8 @@ public class Booking {
     private LocalDateTime dropOff_datetime;
     private long pickup_location_id;
     private long dropOff_location_id;
-    private double estimated_total_price;
-    private double actual_total_price;
+    private BigDecimal estimated_total_price;
+    private BigDecimal actual_total_price;
     private BookingUtils.BookingStatus booking_status;
     private BookingUtils.BookingType booking_type;
     private String notes;
