+++ src/main/java/com/crs/main/model/Rent.java
@@ -26,8 +26,8 @@ public class Rent {
     private LocalDateTime dropOff_datetime;
     private long pickup_location_id;
     private long dropOff_location_id;
-    private double estimated_total_price;
-    private double actual_total_price;
+    private BigDecimal estimated_total_price;
+    private BigDecimal actual_total_price;
 
     private RentUtils.RentType rent_type;
     private PaymentUtils.PaymentPlan payment_plan;
